refactor started

+++ src/main/java/com/javalab/contacts/service/command/DeletePhoneCommand.java
@@ -7,6 +7,8 @@ import com.javalab.contacts.repository.impl.PhoneDtoRepositoryImpl;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import static org.apache.commons.lang3.StringUtils.isNotBlank;
+
 public class DeletePhoneCommand implements Command{
 
     private PhoneDtoRepository repository = new PhoneDtoRepositoryImpl();
@@ -16,8 +18,10 @@ public class DeletePhoneCommand implements Command{
         String[] selectedIds = request.getParameterValues("selectedId");
         if (selectedIds != null) {
             for (String stringId: selectedIds){
-                Integer id = Integer.parseInt(stringId);
-                repository.delete(id);
+                if(isNotBlank(stringId)){
+                    Integer id = Integer.parseInt(stringId);
+                    repository.delete(id);
+                }
             }
         }
     }
